requirements.txt
Flask==1.1.2
gunicorn==20.0.4
jinja2==2.11.3
markupsafe==1.1.1
itsdangerous==1.1.0
Werkzeug==1.0.1
google-cloud-storage==1.42.0
six==1.16.0